+++ evalml/pipelines/component_graph.py
@@ -257,7 +257,7 @@ class ComponentGraph:
 
         parent_inputs = [
             parent_input
-            for parent_input in self.get_parents(self.compute_order[-1])
+            for parent_input in self.get_inputs(self.compute_order[-1])
             if parent_input[-2:] != ".y"
         ]
         for parent in parent_inputs:
@@ -324,7 +324,7 @@ class ComponentGraph:
             x_inputs = []
             y_input = None
 
-            for parent_input in self.get_parents(component_name):
+            for parent_input in self.get_inputs(component_name):
                 if parent_input[-2:] == ".y" or parent_input == "y":
                     if y_input is not None:
                         raise ValueError(
@@ -503,14 +503,14 @@ class ComponentGraph:
             if isinstance(component_class, Estimator)
         ]
 
-    def get_parents(self, component_name):
-        """Finds all of the inputs for a given component, including the names of all parent nodes of the given component
+    def get_inputs(self, component_name):
+        """Retrieves all inputs for a given component.
 
         Arguments:
-            component_name (str): Name of the child component to look up
+            component_name (str): Name of the component to look up.
 
         Returns:
-            list[str]: List of inputs to use
+            list[str]: List of inputs for the component to use.
         """
         try:
             component_info = self.component_dict[component_name]
@@ -667,7 +667,7 @@ class ComponentGraph:
 
     def _get_parent_y(self, component_name):
         """Helper for inverse_transform method."""
-        parents = self.get_parents(component_name)
+        parents = self.get_inputs(component_name)
         return next(iter(p[:-2] for p in parents if ".y" in p), None)
 
     def inverse_transform(self, y):